:building_construction: [#14] moved the information category api code the metadata app
