+++ src/woo_publications/api/urls.py
@@ -4,7 +4,7 @@ from django.views.generic import RedirectView
 from drf_spectacular.views import SpectacularJSONAPIView, SpectacularRedocView
 from rest_framework import routers
 
-from .views.category import InformationCategoryViewset
+from woo_publications.metadata.api.views.category import InformationCategoryViewset
 
 app_name = "api"
