+++ src/woo_publications/metadata/api/views/category.py
